pyproject.toml
[tool.poetry]
name = "drive-backend"
version = "0.1.0"
description = ""
authors = ["lucassing <singlucasmartin@gmail.com>"]

[tool.poetry.dependencies]
python = "3.10.5"
Django = "4.0.5"
djangorestframework = "3.13.1"
psycopg2 = "2.9.3"
daphne = "3.0.2"
whitenoise = "6.2.0"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
python_files = [
    "tests.py",
    "test_*.py",
    "*_tests.py"
]